requirements.txt
numpy==1.16.6
tensorflow==1.15.0
pyyaml==3.13
